refactor: read the scoped credential from the security context, not a parameter
`checkProjectPermission` took an optional `credential` that defaulted to the
one in the security context, so an off-request-thread caller could seem to
supply it directly. It could not: the check delegates to
`checkProjectPermissionNoApiKey`, which reads the credential from the context
regardless, and without one there it takes the admin and supporter bypasses —
so a caller that passed the credential but no context got a *more* permissive
answer than one that set the context up properly.

Both callers already needed the context for that reason, so the parameter was
redundant where it worked and misleading where it did not. It is gone, and
`withSecurityContext` moves next to `TolgeeAuthentication` so both callers
share one copy.

`QaCheckPreviewWebSocketHandler` gains the wrapper it was missing. It accepts
a JWT alone today, which is never a scoped credential, so nothing changes —
but the day it accepts a project API key or an OAuth token, the check would
have silently skipped scope narrowing.
